Rename summary files to improve dx.
