+++ python_components/evaluation/evaluation/exception/__init__.py
@@ -1,6 +1,10 @@
 __all__ = [
+    "ceq_score",
+    "ceq_schema",
+    "ceq_template",
     "deterministic_score",
     "evaluation",
+    "faithfulness_score",
 ]
 
 from evaluation.exception.evaluation import EvaluationWrapper
